requirements.txt
streamlit
PyPDF2
faiss-cpu
requests
numpy
pdfplumber
